@@ -2,23 +2,36 @@
   "name": "christmas-shop",
   "version": "1.0.0",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "build": "webpack",
-    "dev": "webpack serve --open --env development"
+    "dev": "webpack serve --open --env development",
+    "lint": "eslint . --ext .js"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "description": "",
   "devDependencies": {
-    "@awmottaz/prettier-plugin-void-html": "^1.6.1",
-    "copy-webpack-plugin": "^12.0.2",
     "css-loader": "^7.1.2",
+    "eslint": "^8.57.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-webpack-plugin": "^4.2.0",
+    "globals": "^15.13.0",
     "html-webpack-plugin": "^5.6.3",
-    "mini-css-extract-plugin": "^2.9.1",
-    "sass": "^1.80.5",
-    "sass-loader": "^16.0.2",
-    "webpack": "^5.95.0",
+    "mini-css-extract-plugin": "^2.9.2",
+    "prettier": "^3.4.2",
+    "sass": "^1.82.0",
+    "sass-loader": "^16.0.4",
+    "stylelint": "^16.11.0",
+    "stylelint-config-clean-order": "^6.1.0",
+    "stylelint-config-css-modules": "^4.4.0",
+    "stylelint-config-standard-scss": "^14.0.0",
+    "svg-sprite-loader": "^6.0.11",
+    "webpack": "^5.97.1",
     "webpack-cli": "^5.1.4",
     "webpack-dev-server": "^5.1.0"
   }
